quick fix to moisture sensor code

+++ src/include/MoistureSensors.h
@@ -12,8 +12,8 @@ class SoilMoistureSensor
 {
     public:
         explicit SoilMoistureSensor(int pin, int adc_input);
-        void     init();
-        uint16_t read_moisture();
+        void  init();
+        float read_moisture(); // Returns moisture as a percentage (0.0 = dry, 100.0 = wet)
     private:
         int m_adc_channel;
         int m_pin_number;
